backend/requirements.txt
fastapi
uvicorn
requests
pydantic
tenacity
python-dotenv
pytest
